requirements.txt
wheel
fastapi
uvicorn
Pillow
python-multipart
ultralytics